Complete Backend Project

+++ Film-Rental-Store-Sprint/src/main/java/com/cg/model/Category.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Objects;
 
 @Entity
@@ -12,12 +13,15 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id", columnDefinition = "TINYINT UNSIGNED")
-    private Short categoryId;
+    private Integer categoryId;
 
     @NotBlank(message = "Category name is required")
     @Size(max = 25, message = "Category name must not exceed 25 characters")
     @Column(name = "name", nullable = false, length = 25)
     private String name;
+    
+//    @OneToMany(mappedBy = "category", fetch = FetchType.LAZY)
+//    private List<Film_Category> filmCategories; 
 
     @Column(name = "last_update", nullable = false, 
             columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP")
@@ -28,14 +32,14 @@ public class Category {
     }
 
     // All-args constructor
-    public Category(Short categoryId, String name, LocalDateTime lastUpdate) {
+    public Category(Integer categoryId, String name, LocalDateTime lastUpdate) {
         this.categoryId = categoryId;
         this.name = name;
         this.lastUpdate = lastUpdate;
     }
 
     // Getters
-    public Short getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
 
@@ -48,13 +52,21 @@ public class Category {
     }
 
     // Setters
-    public void setCategoryId(Short categoryId) {
+    public void setCategoryId(Integer categoryId) {
         this.categoryId = categoryId;
     }
 
     public void setName(String name) {
         this.name = name;
     }
+    
+//    public List<Film_Category> getFilmCategories() {
+//        return filmCategories; // Return list of Film_Category
+//    }
+//
+//    public void setFilmCategories(List<Film_Category> filmCategories) {
+//        this.filmCategories = filmCategories; // Set list of Film_Category
+//    }
 
     public void setLastUpdate(LocalDateTime lastUpdate) {
         this.lastUpdate = lastUpdate;
